Initial ProgPoW + Stratum support (#115)
* Initial ProgPoW + Stratum support

* Cache for already-calculated ProgPoW hashes

* Fix build

* Added testnet_progpow network + additional pre-validation of potential block headers to avoid unnecessary ProgPoW load

* Only do timestamp block header prevalidation if ProgPoW hashes would be triggered

* Updated block header timestamp+blockheight validation, tightened constraints, added initial grace period, added block height limit

* Update TestNetProgPoW Starting Bitcoin Block
Add BouncyCastle Security Provider to VPM
Add default TestNetProgPoW Config to VPM

* Updated ProgPoW Hash Evaluation Timestamp
Restructured ProgPoW Header Plausibility Filter Flow

* Limits on DAG caching
Pruning of ProgPoW Hash Cache
Tightened allowances on timestamp drift from expected for block header timestamp validations

* Merge fixes

* ProgPow fork height accessible from network parameters

* removed TODO comment that no longer applies

* Moved 30-second blocktime and start-time epoch to network parameters

* Updated progpow test btc block

* Updated ProgPoW to 0.9.3 parameters
ProgPoW seed reduced to 62 (effective) bits
Additional keccakF800 run added to seed generation
Reordererd merge and math portions of ProgPoW for uniqueness
Updated EPOCH length to 8000 and EPOCH_OFFSET to 332

* Updated second seed Keccak round inputs

* Updated ProgPoW with fixes for EPOCH offset

* Progpow header size (#176)

* Updated block header and nonce size

* masking fix

* removed progpow fork height from testnet defaults

* version bump

* typo fix

* Updated nonce for ProgPoW calculation to reflect 40 bits in hashBlock(), added new NONCE_MASK to handle post-vBlake nonces with appropriate shift to preserve all 40 bits during block header assembly

* Updated block header chopping for ProgPoW calculation to account for 5-byte nonce.

* Fixed nonce assembly into block header in assembleBlockHeader()

* Added byteArrayToLong conversion

* Updated TestNetProgPoW starting Bitcoin block.

* Network checks configurability

* Fixes for header changes

* Fixed Long nonce extraction from ProgPoW block header

* Updated synthetic extra nonce to allow 40-bit nonces.

* 8 -> 6 hex characters for SYNTHETIC_EXTRA_NONCE

* 13 extra Keccak rounds for ProgPoW loop

* Updated ProgPoW Lanes to 32

* PROGPOW_LANES=15

* DAG_LOADS=8

* HASH_BYTES = 128

* Reduced seed size to 55 bits, updated PROGPOW_CNT_DAG to 128, PROGPOW_CNT_MATH to 20

* Updated Keccak THETA function

* Modified Keccakf800 algorithm

* Cleanup

* Updated Bitcoin starting block for ProgPoWTestNet

* merge fix

+++ veriblock-core/src/main/java/org/veriblock/sdk/models/VeriBlockBlock.kt
@@ -21,7 +21,7 @@ open class VeriBlockBlock(
     merkleRoot: Sha256Hash,
     timestamp: Int,
     difficulty: Int,
-    nonce: Int
+    nonce: Long
 ) {
     val raw: ByteArray
     val height: Int
@@ -32,7 +32,7 @@ open class VeriBlockBlock(
     val merkleRoot: Sha256Hash
     val timestamp: Int
     val difficulty: Int
-    val nonce: Int
+    val nonce: Long
     val hash: VBlakeHash
 
     init {
